requirements.txt
azure-kusto-data==3.1.1
azure-kusto-ingest==3.1.1
requests~=2.27.1
